refactor(flux): move stuff around inside package
