+++ cmd/daemon/flux/exporter.go
@@ -1,0 +1,11 @@
+package flux
+
+import (
+	"context"
+)
+
+// Exporter sends a formatted event to an upstream.
+type Exporter interface {
+	// Send a message through the exporter.
+	Send(c context.Context, message Message) error
+}
